# api-gateway/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GIGA API Gateway",
    "version": "2.0.0",
...
    "contact": {
      "name": "GIGA Platform Team",
      "email": "api@giga.com"
    }
  },
  "servers": [
    {
      "url": "https://giga-giga-production.up.railway.app",
      "description": "Production API Gateway"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Supabase Auth"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/social/stories": {
      "post": {
        "summary": "Create a story",
        "description": "Create a new story (expires in 24h)",
        "tags": ["Social Media"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["media_url"],
                "properties": {
                  "media_url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "caption": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Story created"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get stories",
        "description": "Get stories from friends",
        "tags": ["Social Media"],
        "responses": {
          "200": {
            "description": "List of stories"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/social/stories/{id}/viewers": {
      "get": {
        "summary": "Get story viewers",
        "description": "Get list of users who viewed a story",
        "tags": ["Social Media"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Story viewers"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  }
}